teste

+++ server/src/routes/memories.ts
@@ -1,12 +1,20 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
 import { z } from 'zod'
-import { request } from 'http'
 
 export async function memoreisRoutes(app: FastifyInstance) {
   /* --------------------------------------------------------------- */
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify() // garante que o token esta vindo do front
+  })
+  /* --------------------------------------------------------------- */
+  app.get('/memories', async (request) => {
+    // console.log(request.user.sub)
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -14,13 +22,13 @@ export async function memoreisRoutes(app: FastifyInstance) {
     return memories.map((memory) => {
       return {
         id: memory.id,
-        coverUrl: memory.converUrl,
+        coverUrl: memory.coverUrl,
         excerpt: memory.content.substring(0, 115).concat('...'),
       }
     })
   })
   /* --------------------------------------------------------------- */
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -30,29 +38,33 @@ export async function memoreisRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
   /* --------------------------------------------------------------- */
   app.post('/memories', async (request) => {
     const bodySchema = z.object({
       content: z.string(),
-      converUrl: z.string(),
+      coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
     })
-    const { content, isPublic, converUrl } = bodySchema.parse(request.body)
+    const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
     const memory = await prisma.memory.create({
       data: {
         content,
-        converUrl,
+        coverUrl,
         isPublic,
-        userId: '4445291c-fd56-4208-9c93-7420dbc04e79',
+        userId: request.user.sub,
       },
     })
     return memory
   })
-
   /* --------------------------------------------------------------- */
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -61,29 +73,48 @@ export async function memoreisRoutes(app: FastifyInstance) {
 
     const bodySchema = z.object({
       content: z.string(),
-      converUrl: z.string(),
+      coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
     })
-    const { content, isPublic, converUrl } = bodySchema.parse(request.body)
+    const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
       data: {
         content,
-        converUrl,
+        coverUrl,
         isPublic,
       },
     })
     return memory
   })
   /* --------------------------------------------------------------- */
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     await prisma.memory.delete({
       where: {
         id,
